move createGame into own handler, move handlers into own package
