+++ server/src/main/java/server/handlers/LoginHandler.java
@@ -1,4 +1,4 @@
-package server;
+package server.handlers;
 
 import dataaccess.*;
 import model.*;
