+++ server/src/main/java/server/handlers/RegisterHandler.java
@@ -1,4 +1,4 @@
-package server;
+package server.handlers;
 
 import dataaccess.*;
 import model.*;
